Swift - Updated MDSDocumentStorage associationGetIntegerValue to return Int64 instead of Int. Swift - Updated MDSSQLite implementation to leverage SQLite sum.

+++ Source/Swift/MDSDocumentStorage.swift
@@ -87,7 +87,7 @@ public protocol MDSDocumentStorage {
 	func associationIterate(for name :String, to toDocumentID :String, fromDocumentType :String,
 			proc :(_ document :MDSDocument) -> Void) throws
 	func associationGetIntegerValue(for name :String, action :MDSAssociation.GetIntegerValueAction,
-			fromDocumentID :String, cacheName :String, cachedValueName :String) throws -> Int
+			fromDocumentID :String, cacheName :String, cachedValueName :String) throws -> Int64
 
 	func cacheRegister(name :String, documentType :String, relevantProperties :[String],
 			valueInfos :[(name :String, valueType :MDSValueType, selector :String, proc :MDSDocument.ValueProc)]) throws
@@ -232,7 +232,7 @@ extension MDSDocumentStorage {
 	//------------------------------------------------------------------------------------------------------------------
 	public func associationGetIntegerValue<T : MDSDocument>(from document :T,
 			action :MDSAssociation.GetIntegerValueAction, toDocumentType :String, cachedValueName :String) throws ->
-					Int {
+					Int64 {
 		// Return value
 		return try associationGetIntegerValue(
 				for: associationName(fromDocumentType: T.documentType, toDocumentType: toDocumentType), action: action,
@@ -242,7 +242,7 @@ extension MDSDocumentStorage {
 	//------------------------------------------------------------------------------------------------------------------
 	public func associationGetIntegerValue<T : MDSDocument>(for name :String,
 			action :MDSAssociation.GetIntegerValueAction, from document :T, cacheName :String? = nil,
-			cachedValueName :String) throws -> Int {
+			cachedValueName :String) throws -> Int64 {
 		// Return value
 		return try associationGetIntegerValue(for: name, action: action, fromDocumentID: document.id,
 				cacheName: cacheName ?? T.documentType, cachedValueName: cachedValueName)
